tests: move to file with other _explanation tests
